feat: Add export to Excel functionality for clients/suppliers
This commit adds the ability to export client and supplier data
to Excel. This feature is implemented by adding an "Export to
Excel" option in the action menu for both the client and
supplier index pages. The export functionality is handled by the
`exportExcel` method in the respective index actions.
Also add bills_count translation in language files.